Suppress deprecation warnings in html5lib
html5lib currently issues a DeprecationWarning in python ^3.7. This
warning is fixed here:

https://github.com/html5lib/html5lib-python/commit/4f9235752cea29c5a31721440578b430823a1e69

I have spoken with the maintainers (#whatwg on freenode) and they said
that a new release of html5lib is not scheduled as the package is not
actively maintained, but it sounded like they were aware of the issue
and could release a new version. (I did not press further.)

This is a very minor change with no major intended side effects, other
than that it cleans up output while running tests (and presumably from
any clients using the legacy_repository type).