Fix request authentication when credentials are included in URLs (#751)
* Fix authentication failure for some APIs

Fixes an issue which occurs when APIs return authentication information in the URLs embedded in JSON responses.

* Test authentication on same host with duplicated credentials

Add test to ensure that basic authentication credentials in request URLs does not break authentication when they credentials match. This test purposefully does not examine program behavior when URL credentials and session credentials are different.