Allow to manage project version with git tags
To avoid to manage textual version of a project allow poetry
to retrieve those informations directly from git tags.

To determine the current version number we currently retrieve
the highest available tag number.

If the version is not available in the pyproject.toml then
poetry try to retrieve version from git so project can continue
to manage these informations manually.

If the version is not available in the pyproject.toml and if
it's not a git repository then an exception is raised and we
retrieve the previous behavior of poetry.