get-poetry.py: Use correct python when installed with non-default python
On CentOS 6, `/usr/bin/env python` is Python 2.6. This results in poetry
being unusable since 2.7 is required. Poetry should still work when an
alternative Python is installed. That is, you should be able to run
`python2.7 get-poetry.py` and end up with a usable installation of
poetry.

This will use the python executable used to run get-poetry.py in the
shebang, ensuring that a valid python is used.