Canonicalise virtual environment name
poetry's in-memory package representation saves the package name in
a canonicalized form (replace underscores with dashes).

When virtual environments are created this form is used instead of the
actual name, while operations on the virtual environment are still
performed with the base environment name. This fails if the environment
name contains a character, that would normally be replaced during
canonicalisation.