Detect Conda when looking for a virtual env
Currently Poetry will detect that it is running in a usual virtual env
created with "python -m venv" and will not create an additional env.

This commit extends this logic to Conda, which uses different
environment variables to indicate running in a virtual env.

Conda's "base" env is treated specially to avoid polluting global
namespace.