When Building Wheels, Resolve Relative Path Dependencies Correctly
Building wheels unpacks the sdist which includes a copy of the
`pyproject.toml`.  If dependencies are delcared using relative paths
(such as `mymodule = { path = '../mymodule' }`), these relative paths
are not available from the temporary directory created to unpack and
build the wheel.  This fix retains the original path to the project for
those relative dependencies.

Fixes #266.