Add support to use repository url when unspecified
Currently, even if a url is defined for a user level configured
repository we require it also be specified in the project's
pyproject.toml. With this change, if no url is specified in a
pyproject.toml a user level configured repository url is used as a
fallback. This also improves the error message when no URL is found for
a repository.