Relax licence restrictions to support custom licences
This change adds support of use of custom licenses for projects. If the
value specified for `licence` is not in the SPDX License List or
"Proprietary", we treat it as a custom licence withe value (as is) as
the name.

As a side-effect, the validation of licence is no longer performed when
executing `poetry check`.

Resolves: #2020