feat(installer): add only dev mode option
This commandline parameter adds the ability to install only dev
packages. That is often needed, when poetry is run in a CI/CD system
like GitLab CI or GitHub actions, where often not the whole test
suite for every PR is used, but only linters or static
type checkers are required. These often reside in dev dependencies.