fix: fix handling of packages referenced by the multiple 'extra'
close #2676

For a package referenced by a single extra, a SingleMarker is used, where the result of without_extra is AnyMarker.
In the case of a package referenced by a multiple extra, a MarkerUnion is used, where the result of without_extra is an empty MarkerUnion.
There was a problem because the package was not installed if the package's Marker was an empty MarkerUnion.