Change usage example for poetry export
The existing example:
```bash
poetry export -f requirements.txt > requirements.txt
```
The problem with this is that any additional output from poetry will also end up in the requirements file. For instance, if the `poetry.lock` file is not present, poetry will output to stdout some extra text telling you about this. This breaks the requirements file. (I was burned by this problem, hence this PR)

Seeing as there is already a documented option to specify the filename directly, which resolves this issue, updating the docs like this seems reasonable.