Teach `poetry run` to find `.cmd` scripts on Windows
The Poetry script generation creates `.cmd` scripts, not `.exe` files,
but `poetry run` did not know to look for `.cmd` scripts resulting in
a `FileNotFoundError` when running custom scripts on Windows.

This doesn’t reproduce with `poetry shell` because the executable search
happens in the shell (like PowerShell).

Because this function now needs to search across two suffixes on
Windows (and this may increase in the future due to Windows’ use of file
extensions to mark a file executable), it was refactored and the
negative logic flipped to read more clearly:

1. Get the path to the bin in Poetry’s environment
2. On Windows, test if path.(exe, cmd) exists, and return it (handling
   the existing edge case with base paths).
3. If the path (without an extension) exists, return it
4. Return the original ‘bin’ if none of the above succeeded