check archive hash only for matching filename
Previously a downloaded package was checked against all available hashes.
This is now changed to only allow the hash from the lock file matching the
current archive name.

The user experience benefits with more precise error messages:
The use can now differenciate if the hash does not match
or if the file is not present in the lock file.