refactor: Add "mirror" param for LegacyRepository
... to enable further improvements on mirrored repository.

One example is that the cache dir of the mirrored PyPI repository should
also be named "pypi". However, the repository name "pypi" is reserved
for the original PyPI. Doing this change to LegacyRepository will not
only allow this, but also enable other developers to make changes if
mirrored LegacyRepository is broken.

The corresponding tests are also added. Note that there are some changes
on the paths to some files.