Handle RepositoryError on secondary repos
If find_packages() for any pool repository raised a RepositoryError, no
results would be returned even if the package had already been located
in the default repository. This is problematic if a secondary returns
a 403, such as in the case of Artifactory pypi repositories which can
return 403 when a package does not exist.

This change permits find_packages to continue in the event that
a request to a secondary repo raises a RepositoryError.