Make the dependency on poetry-core more restrictive
Since `poetry-core` will see some changes in order to modernize its codebase it's better to be prudent here by restricting it to `~1.0.2` for the `1.1` branch of Poetry until we can be sure we can loosen it.

Note that it does change anything regarding the constraint in the `build-system` section since the high level API is unlikely to change for the foreseeable future.