Run tests involving file fixtures in a temporary directory
A major barrier towards running tests in parallel is that some of them
modify files in project fixtures. As a simple fix, copy the whole
project fixture directory tree to a temporary folder and run tests
there.

Some of the fixtures for installer tests contain paths to the project
fixtures. To make those work, now that the paths are temporary, make
installer tests use the project fixture directory as their working
directory.