Fix `remove` command to handle `.venv` dirs
When the `virtualenvs.in-project` config option was set to `true`, the
Python virtual environment is created in a directory named `.venv`,
local to the project root.

After running `poetry install`, the env directory is listed like this:

    $ poetry env list
    .venv (Activated)

However, the `env remove` command raised multiple exceptions when
trying to remove this environment:

    $ poetry env remove .venv
    /bin/sh: .venv: command not found

    EnvCommandError

    [...]

It took quite a bit of time to get to the bottom of why this was
happening, but I eventually found the bug.

When support was added for local `.venv` directories, several changes
were made throughout the project to special-case the handling of this
environment.

However, the `EnvManager.remove()` method contained no special-casing
for the `base_env_name` variable, meaning that the
`python.startswith(base_env_name)` check never passes, and as a result,
there is a cascade of errors.

This patch fixes the bug by adding special-casing for the
`base_env_name` variable.