Fix install on MingW on Windows
python, when installed via (msys) MingW on Windows, can be found under a path
structure similar to that you might find on *nix systems.

The install-poetry.py script was failing to take this into consideration,
causing installation issues winth Windows + MingW.

A similar fault in poetry itself was resolved in #3713.