Fix Install Error Not Shown (#4474)
An AttributeError in the install script happens when an install
operation failure doesn't result in stderr output. In these cases, it
appears the data is written to stdout instead. To fix this, the error
output will use stdout when no stderr output exists. If neither of these
exist a "No output available." message will be shown instead to prevent
the AttributeError from obscuring what is actually going on in the flow
of the install script.