fix: Don't try to parse `Link`s as `Path`s
Previously, `Link`s and `Path`s were passed in to `pip_install` as a
string. This makes it hard for pip_install to parse out the extension of
the file to install. By accepting only `Link`s or `Path`s, `pip_install`
can parse out the extension using the `Link`.