Improves hash mismatch exception language (#4573)
* Improves wording of exception when calculated hash absent from lockfile and
extracts archive hashing and validation to a separate method

I think this could be further improved by representing the hashes as
a dataclass containing the hash type and the string digest. The types of
hashes would be more robust then instead of hardcoding sha256 here and
in `FileDependency.hash()`. It would also enable enumeration of
hash types of known hashes so that `FileDependency.hash()` could return
a list of hashes based on requested types.

Doing this might alleviate the md5/sha256 problems arising from changes
in poetry-core 1.0.6+ that affect Poetry 1.1.7+ users using Artifactory.

Co-authored-by: Arun Babu Neelicattu <arun.neelicattu@gmail.com>