ci: add clarity to pytest
This provides greatly improved diffs when running tests locally. Clarity
is disabled in CI as it doesn't work well with the simple log viewers
there.

Also note that Pytest is now run in (very) verbose mode by default. This
is because Clarity requires it, and also because it is much more useful
feedback in CI.