command/add: introduce dependency sorting
python-poetry/poetry#3996

When installing a package with `poetry add`, the package is appended to
the appropriate section of the pyproject.toml. Some developers have
dependencies sorted alphabetically, so they have to manually sort the
new dependency into the appropriate place in pyproject.toml.

This commit will add a poetry config setting `dependencies.sort`, which
can also be configured with the corresponding environment variable
`POETRY_DEPENDENCIES_SORT`. If `true`, dependencies will be sorted
alphabetically when adding, keeping `python` at the top of the list.

Initially, the setting was configured with a `--sorted (-S)` option, but
it is easier to have a config setting instead of repeatedly having to
enter the `--sorted` option.

Poetry dependency groups (added in 1.2.0a2) will be included and tested.