ensure setup.py editable builds do not use pep517
Prior to this change when Poetry builds a project with a setup.py file,
as in the case when `generate-setup-file` was set to true, `pip`
fallback was invoked with a PEP 517 isolated build causing it to fail.

This was a regression in functionality from Poetry 1.1 build script
usage. However, note that build script usage is an experimental
feature.