ci: update to Ubuntu 22.04 and macOS 12
Testing against newer versions will help us keep abreast of any upcoming
issues. The older versions can be added as partial matrix entries if
desirable, but I do not believe this to be necessary.

Also re-order mypy in the workflow so that type errors fail fast (before
tests are run).