Correct URL in test_parse_dependency_specification
The URL path is used by `mock_download` to determine the path relative
to the fixtures directory to get the download result.

I believe this has been passing on CI because the `copy_or_symlink` call
in `mock_download` happily creates a symlink to a non-existent path, but
when running on Windows without permission to create symlinks,
`shutil.copyfile` is called as a fallback, and fails trying to copy a
non-existent file.