fix: add missing deprecation warning for `--dev` (#6475)
`--dev` is marked as deprecated for both
[add](https://python-poetry.org/docs/cli/#options-4) and
[remove](https://python-poetry.org/docs/cli/#options-5) commands in the
documentation and emit a warning when used, but this is not reflected in
the CLI help messages.

I also took the occasion to update the documentation to suggest
replacements. Once backported, this would also require an update of the
1.2.0 blog post announcement, which doesn't mention those deprecations.

(cherry picked from commit 844a720683b6c53bd8be05402c4e79909df5ee36)