ci: use flake8-pie (#6164)
Relates-to: #4776 

Adding the hook leads to four new warnings for the repo.
-
[PIE803](https://github.com/sbdchd/flake8-pie#pie803-prefer-logging-interpolation):
prefer-logging-interpolation
This check produces false positives (`debug()` calls are always flagged
as if they belong to a logger). PR ignores such instances and fixes the
rest.
-
[PIE786](https://github.com/sbdchd/flake8-pie#pie786-precise-exception-handlers):
precise-exception-handlers
PR ignores instances where the intention is indeed to catch any
exceptions and fixes the rest.
-
[PIE798](https://github.com/sbdchd/flake8-pie#pie798-no-unnecessary-class):
no-unnecessary-class
All instances are ignored via an additional entry in flake8 config.