Adds 'self source add|remove|show' for managing Poetry plugin sources
This reuses the existing command infra provided by the `source` command,
using the SelfCommand infra to change which pyproject.toml is the
target.

I tried to write some tests for this, but realized that the tests would
be modifying the developer's valued Poetry system pyproject.toml.
I think it might be OK to rely on the existing source add, remove, and
show tests, but I'll share my test code in a comment on the PR for this
branch.