Clarifies the dev-dependencies -> group.dev.dependencies switch
As written, it seemed to prioritize the older section notation without clarifying that the older syntax still works but shouldn't be used. This reorder visually prioritizes the preferred syntax as of 1.2.x while clearly explaining what the old syntax is during this transition period.

The best next step here would be a clear definition of the version in which dev-dependencies will go away.