Clarify dots canonicalisation change is breaking
The handling of package names within the Python packaging ecosystem
is unfortunately not in a great state at the moment. There exist
a large number of packages which contain characters which (some of)
the specs indicate are invalid, yet there is not a clear migration
path nor destination at the moment.

Poetry's approach is to normalise project names towards their
"canonical" form, meaning that existing projects on PyPI which
use the more relaxed name forms will be silently renamed if they
move to (recent versions of) Poetry. As well as changing the name
of these packages (which maintainers are unlikely to expect nor
desire) this unfortunately ends up breaking introspection via
`importlib.metadata` in Python 3.8 and 3.9 which are not aware of
the canonicalisation rules which Poetry is using.

Adding this compatibility note informs maintainers so that they can
decide how (and if) they want this name canonicalisation to happen.

See also https://github.com/python-poetry/poetry/issues/6198