fix: fix support for packaging v20.4
Poetry v1.2.2 does not work with packaging v20.4
(which is an allowed version of packaging in the pyproject.toml file).

This is because the `from packaging.utils import NormalizedName` line
before [packaging v20.5][1]
only works if `typing.TYPE_CHECKING` is True.

This is not an issue on the `master` branch, since there this
import is already hidden behind an `if TYPE_CHECKING:` guard.
However, the v1.2 branch never cherry-picked the appropriate
commit adding this change.

[1]: https://github.com/pypa/packaging/commit/eca30e0435f99084682d529d997f5c5fe1cd43fb

Fixes: https://github.com/python-poetry/poetry/commit/04c5cbf043610dd5d89b7d980bf8a137948c4fa5