Modify 'Usecase 2' for poetry+tox
This is a longer config, but it better demonstrates how
`poetry install --no-root --sync` can be used to sync locked
dependencies to tox environments.

Importantly, it demonstrates the following improvements over the
existing example:
- it does not run commands with `poetry run`, which *looks* correct
  but actually means that the `commands` are executing from the poetry
  environment, rather than the tox environment
- it shows the application of dependency groups to tox environments
- it shows differentiated use of `skip_install` (and thereby
  encourages correct usage in concert with `mypy` in particular, which
  many projects struggle to handle)

The main downside of this new example is that it is longer and more
complex for the very simple cases.

However, if we take "Usecase 2" as describing an application (rather
than a library) which wants a developer experience in which locked
dependencies are used for linting and test steps, then this is a more
complete and realistic sample config.