Update tox example 2 to set install_command
As suggested by @finswimmer, use `install_command = ...` to pass
`--no-deps` to pip in the second tox usage example. This skips an
unnecessary step in which pip installs dependencies which poetry will
then handle in the sync step.

Additionally, normalize the style of commands a bit in the proposed
update to Usecase 2. Commands are no longer inlined with the config
key, which makes the example slightly longer, but also more consistent
with the other examples.