Consider system packages as installed if the venv includes them
This changesets adds a getter to `env.virtual_env.VirtualEnv` to check
whether it has access to system packages, and overrides the
`is_path_relative_to_lib` method to take it into account.

This will prevent Poetry from reinstalling system packages in the venv
when they are already installed with a compatible version.

Fixes: #6035

Signed-off-by: Aurélien Bompard <aurelien@bompard.org>