log: smarter handling of third-party logging
It turns out the logger name might not be enough for figuring out which third-party package is currently logging, because nothing guarantees such package even uses named loggers (or a sane naming convention).

As the LogRecord lacks the name of the package inside a module, this change adds logic that parses it out of the LogRecord's pathname.