Update description of --why flag for poetry show
Changes:
- Update description in `docs/cli.md`
- Kept the description in the `docs` in sync with the `cli` by updating
`src/poetry/console/commands/show.py`

Motivation:

The current description was unclear. "display why a package is included"
could suggest that users could get back a motivation or rationale behind
why the package is added. But that's not the case.

`--why` returns whether a package is a direct dependency or whether it
is itself a dependency of another package in the project.

Also see:  https://github.com/python-poetry/poetry/issues/8816