ci/tests: add smoke-test step
This step verifies that an external tool can build the project, that
the distfiles have valid metadata, and that the bdist can be installed.

We then perform a smoke test, ensuring that the version of the
(third-party) built package matches the expected version from
pyproject.toml.

This is meaningful as `poetry build` uses the in-tree copy of
`poetry-core`, and we expect to be build and installed by other tools.