MockDulwichRepo: mimick git behaviour more closely
When git clones a repo in a base directory, it will place the clone
in a subdirectory of the base. That subdirectory is named after
the repo url name, which is the "testing" right before the .git in
https://github.com/test/testing.git

This behaviour may introduce race conditions to Poetry. Let the
mock repo mimic it, so that the (absence of) race conditions
can be verified.