feat(env): prefer embedded pip for poetry use
Presently, when Poetry dispatches pip commands for uninstallation, this
uses the pip version installed in the active virtual environment. This
can potentially cause issues like #10089.

This change, switches to preferring the embedded pip wheel and makes
preferring the in-env version the exception, while preferring it when a
user uses `poetry run pip`.