chore(ci): reduce cirrus usage
FreeBSD testing was initially introduced when Poetry was distributed in
platform specific bundles. This is no longer the case and the value of
these tests have reduced significantly, especially when considering
newer Python versions (>=3.12) are still not available via ports.

This change reduces the jobs to be more of a smoke test than a full
coverage. Mypy tests have been removed along with testing on Python
versions <3.11.