feat(config): installer.build-config-settings.pkg
This change introduces the `installer.build-config-settings.<pkg>`
configuration option to allow for PEP 517 build config settings to be
passed to the respective build backends when a dependency is built
during installation.

This feature was chosen not to be exposed via an addition to the
dependency specification schema as these configurations can differ
between environments.

Resolves: #845