executor: propagate original error even if pkg.to_dependency() fails
this fixes the following regression, introduced somewhere between 1.8.4
and 2.0.1:

a git dependency which fails wheel build does not output the underlying
error that caused the wheel build failure. eg.

	podman run -ti -w /tmp --entrypoint /bin/sh --rm python:3.13.3-alpine3.21 -c "pip install poetry && poetry new . && poetry add 'python-kadmin @ git+https://github.com/authentik-community/python-kadmin.git@v0.2.0'"

the root cause is that the exception handling in _execute_operation()
calls pkg.to_dependency(), which fails like:

	ValueError: Invalid git url "/root/.cache/pypoetry/virtualenvs/tmp-6WcazSRI-py3.13/src/python-kadmin"

and the underlying build error is therefore never output.